swarmit/server: unify daemon and dashboard into one swarmit-server entry point
AI-assisted: Claude Opus 4.7
